the input now works, and can be set just by having the function name instead of the hard-coded func = DoSomething; All functions that will be bindable will be in a vector, which is not redundant considering they need names and descriptions for UI somewhere anyways This vector is going to be immutable, unlike the keyMap
The next step will be making an actual canvas,
but in terms of the inputManager expansions, it's being able to save and load json files of edited inputs
This is not needed for a minimum sensibly future-proof product though, so I will do it later

+++ SagePaint/SagePaint/app/MainApp.cpp
@@ -73,7 +73,7 @@ void MainApp() {
 	glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 3);
 	glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
 
-	GLFWwindow* window = glfwCreateWindow(Screen_width, Screen_height, "Sage Paint Alpha",NULL, NULL);
+	GLFWwindow* window = glfwCreateWindow((int)Screen_width, (int)Screen_height, "Sage Paint Alpha",NULL, NULL);
 	if (!window)
 	{
 		IDLOG("Window failed to init!")
@@ -96,8 +96,6 @@ void MainApp() {
 
 
 	//'start()'
-	
-	
 	go = std::make_shared<CanvasObject>();
 	
 	ImagePtr images[] = {
@@ -118,8 +116,8 @@ void MainApp() {
 
 		int width, height;
 		glfwGetFramebufferSize(window, &width, &height);
-		Screen_width = width;
-		Screen_height = height;
+		Screen_width = (float)width;
+		Screen_height = (float)height;
 		Screen_ratio = width / (float)height;
 		glViewport(0, 0, width, height);
 		glClear(GL_COLOR_BUFFER_BIT);
@@ -145,15 +143,15 @@ void MainApp() {
 
 		//DLOG(xpos << "x")
 		//DLOG(ypos << "y")
-		go->pos.x = xpos-Screen_width/2;
-		go->pos.y = -ypos+Screen_height/2;
+		go->pos.x = ((float)xpos)-Screen_width/2;
+		go->pos.y = -((float)ypos)+Screen_height/2;
 		if ((float)glfwGetTime() - timestamp > 5.0f) {
 			timestamp = (float)glfwGetTime();
 			i++;
 			i %= 5;
 			go->LoadImageSync(images[i]);
-			go->scale.x = images[i]->width;
-			go->scale.y = images[i]->height;
+			go->scale.x = (float)images[i]->width;
+			go->scale.y = (float)images[i]->height;
 
 		}
